refactor(tests): split Date.test.ts; fix flaky tests and non-standard value warning
- Split 805-line Date.test.ts into four focused files (≤250 lines each):
  Date.fromJson.test.ts, Date.toString.test.ts, Date.zeroYear.test.ts,
  Date.intercalary.test.ts
- Move shared king/eponym fixtures into test-support/date-fixtures.ts
- Fix flaky Corpus.integration.test.ts: replace unawaited fireEvent.click +
  waitForText with container.toHaveTextContent; update 'Without session' snapshot
- Fix flaky DatesInTextSelection 'updates a date in the list': pin first date
  to SE 5.5.5 to prevent random collision with test's edit inputs (day=18,month=10)
